refactor: `NewCardScreen` 컴포저블을 `NewCardScreenActivity` 에서 분리
