+++ app/src/main/java/woowacourse/payments/NewCardScreenActivity.kt
@@ -40,67 +40,3 @@ class NewCardScreenActivity : ComponentActivity() {
         }
     }
 }
-
-@Composable
-fun NewCardScreen() {
-    var cardNumber by remember { mutableStateOf("") }
-    var cardExpiryDate by remember { mutableStateOf("") }
-    var cardHolderName by remember { mutableStateOf("") }
-    var cardPassword by remember { mutableStateOf("") }
-
-    Scaffold(
-        modifier = Modifier.fillMaxSize(),
-        topBar = {
-            NewCardTopBar(
-                onBackClick = { },
-                onSaveClick = { },
-            )
-        }
-    ) { innerPadding ->
-        Column(
-            modifier = Modifier
-                .padding(innerPadding)
-                .padding(horizontal = 24.dp)
-                .fillMaxWidth()
-        ) {
-            Spacer(modifier = Modifier.height(14.dp))
-            PaymentCard(
-                modifier = Modifier
-                    .align(Alignment.CenterHorizontally)
-            )
-
-            Spacer(modifier = Modifier.height(40.dp))
-            CardNumber(
-                value = cardNumber,
-                onValueChange = { cardNumber = it },
-                modifier = Modifier
-                    .fillMaxWidth()
-            )
-
-            Spacer(modifier = Modifier.height(30.dp))
-            CardExpiryDate(
-                value = cardExpiryDate, onValueChange = { cardExpiryDate = it },
-                modifier = Modifier.fillMaxWidth(0.5f)
-            )
-
-            Spacer(modifier = Modifier.height(30.dp))
-            CardHolderName(
-                value = cardHolderName, onValueChange = { cardHolderName = it },
-                modifier = Modifier.fillMaxWidth()
-            )
-
-            Spacer(modifier = Modifier.height(10.dp))
-            CardPassword(
-                value = cardPassword,
-                onValueChange = { cardPassword = it },
-                modifier = Modifier.fillMaxWidth(0.5f)
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun NewCardScreenPreview() {
-    NewCardScreen()
-}
